Use signed ints for NSTag, 32 bit compatibility
MainMenu.xib fails to compile with a 32 bit ibtool, because of:
	IBXMLDecoder: The value (2483028224) for key (NSTag) is too large to
	fit into a 32 bit integer.
This also breaks HexFiend on 32 bit systems when compiled with a 64 bit
ibtool:
	*** Terminating app due to uncaught exception 'NSRangeException',
	reason: '*** -[NSKeyedUnarchiver decodeInt32ForKey:]: value
	(2483028224) for key (NSTag) too large to fit in 32-bit integer'

Replacing these integers with the decimal representations of their
signed interpretation fixes this problem. 64 bit systems will ignore the
leading 4 bytes when casing to NSStringEncoding turning
	0xFFFFFFFF94000100
into
	0x94000100.

See https://trac.macports.org/ticket/35244 where this was originally
reported and debugged.