Don't include Sparkle for Mac App Store configs
Since Sparkle is not used on the Mac App Store, we don't want
to include Sparkle for MAS configs. However Xcode does not
easily allow varying build phases according to the config.

This commit removes Sparkle from the linker build phase, and
adds it back in the OTHER_LDFLAGS build setting, which can vary
with the config.

It also introduces a new script postprocess_for_mas.sh, which
deletes the Sparkle.framework from the Hex Fiend.app bundle.

One complication is that Xcode can no longer find the framework
for Release builds, since the Sparkle framework's configs are named
differently and thus Sparkle.framework and Hex Fiend wind up in
different directories. This is fixed by adding a Copy Files build phase
to copy the built Sparkle framework to a new build products subdirectory,
called built_frameworks.
(It can't be the build products directory itself, because that would
fail when the configs are the same, like for 'Debug'. In that case,
the Sparkle and Hex Fiend build directories are the same, and you can't
copy a file over itself.)

This commit does not attempt to hide the Check For Updates menu item.
That will come in a future commit.