Anchor templates separately from selection, add Jump to Field menu
Templates are now anchored to position 0 by default. The anchor position can be changed by right-clicking a position and selection "Anchor Template at X". When right-clicking a template field there is a "Jump to Field" menu which scrolls to that field. Also fixed remove highlight feature not working if scrolled.