Rearrange the Font menu to make the non-font items noticeable
Before, depending on the screen size and the number of fonts installed, the menu items that didn’t choose a font could be hidden below the bottom of the menu, which made those menu items hard to discover. Now they are discoverable.

I also renamed the placeholder from “Building...” to “Building font list…” and added an outlet so that the placeholder can be moved in the future without having to update the code.