Fix leak of extraArgs array allocated by Tcl_ParseArgsObjv
From https://www.tcl.tk/man/tcl8.6/TclLib/ParseArgs.htm about remObjv:
Pointer to a variable that will hold the array of unprocessed arguments.
Should be NULL if no return of unprocessed arguments is required. If objcPtr
is updated to a non-zero value, the array returned through this must be
deallocated using ckfree.

Not sure why ckfree takes a char* as argument, but we need to cast
to make the compiler happy. Seems to be what other do too.