Add a PEF template for Classic Mac executables.
PEF is the executable format read by Apple's Code Fragment Manager.  This template provides a deep understanding of most of a PEF container.  Containers are usually stored in the data fork of a file, but can be in memory or a resource fork as well.  Sometimes there are multiple containers (example: CarbonLib in OS 9).