Add a PEF template for Classic Mac executables. (#317)
PEF is the executable format read by Apple's Code Fragment Manager in Classic Mac OS, introduced with the shift to the PowerPC.  This template provides a deep understanding of most of a PEF container.  Containers are usually stored in the data fork of a file, but can be in memory or a resource fork as well.  Sometimes there are multiple containers.  A good example using many of the features of PEF is CarbonLib in OS 9.

Not included: a detailed listing of relocation "instructions" or the exported-symbol hash-table.