Add symlink traversal support to loadTemplates (#327)
I have some templates in another folder that I've symlink'd into the Templates directory. Unfortunately `enumeratorAtPath` does not resolve symlinks explicitly. In the event a file is not a template file, I create an `NSURL` to it and try to resolve it as a symlink. In the event the paths are different (e.g., a symlink was resolved) we traverse the resolved path.