Create BTCBlock-extended.tcl
This Template expands and decodes the ScriptSig, ScriptPubKey, and Witness data in each Transaction. The BTCBlock.tcl Template displays only the raw data of each.

This Template can be compute intensive and can be very slow on older machines when expanding the  collapsed TRANSACTIONS Section. The compute burden can be mitigated using configuration variables in the source code of this Template as explained in the header comments in the code. Decoding is optional for each of the three Transaction objects and be eliminated entirely. As a practical matter disabling decoding for all three will result in performance (and result) indistinguishable from that of the BTCBlock.tcl Template.