Set SKIP_INSTALL=YES for HexFiend_Framework(_iOS) targets (#446)
If SKIP_INSTALL is left to NO, and you try to include HexFiend_Framework as a subproject to another app, and you try to Archive that app in Xcode, Xcode will fail to Archive the app with an error about a missing module.modulemap

This may be recent due to HexFiend framework now using Swift source files.

Since HexFiend framework isn't "installed" on a system but rather included as a dependency to other projects, this change should be okay. The Framework Only (Debug/Release) schemes can still be used when building just the frameworks themselves in isolation.