Template: 6502 CPU instruction set disassembler and Commodore 64 PRG executable format (#468)
This is a template-based 6502 CPU instruction set disassembler intended to assist with inspecting, reverse-engineering and documenting binary formats that contain 6502 code.

Included is a general-purpose template "6502", which disassembles all bytes of a given file. This is useful for identifying which parts of unknown binaries contain 6502 code versus data/images or something else. How this works is detailed in the file comments.

The disassembler itself is a separate "# .hidden = true;" procedure-only template. The intention is to include it in other templates that themselves implement the particular 6502 code encapsulation formats; NES ROM's, some disk images, emulator snapshots, etc.

A Commodore 64 "PRG" template is provided here - an example of one such common format.

(Note: the disassembler was built for usage on binaries typical of 6502 computers, ie, on the order of a few tens of KB. Large binaries might take some time to disassemble.)